updated
splited controllers by functions
