+++ controllers/index.js
@@ -1,5 +1,16 @@
-const contactControllers = require("./contactControllers")
+// const contactControllers = require("./contactControllers")
+const getContacts = require("./getContacts");
+const getContactById = require("./getContactById");
+const deleteContact = require("./deleteContact");
+const createContact = require("./createContact");
+const updateContact = require("./updateContact");
+const updateContactByStatus = require("./updateContactByStatus")
 
 module.exports = {
-    contactControllers
+    getContacts,
+    getContactById,
+    deleteContact,
+    createContact,
+    updateContact,
+    updateContactByStatus,
 }
